Update 27047561-9d76-b37d-100d-1c58e6edf494.md
The statement that a user defined function marked volatile is only 'recalculated whenever calculation occurs in any cells on the worksheet on which this function appears' was just plain wrong. In fact, it will not only recalculate whenever any value in the workbook is changed (calculation cycle or no calculation cycle) but in the case of multiple workbooks open within a single application instance, it will recalculate when values change on other workbooks open within the same application instance.