Update 764dafed-d4e3-82b9-df8c-68a358319491.md
In Office 2010, the name of the argument that specifies whether the table data has headers is "HasHeaders".  Given that in 2013, the signature for this method (as shown on this page) does not include a TableStyleName argument, I would assume that the proper name for this argument is "HasHeaders".

Also, since I don't have Office 2013, I can not verify whether the TableStyleName argument has been removed.  If not, then it should be added to this page (with a proper description of its use).