Update d26e6786-064a-174c-5b9f-79e85b34f59b.md
Passing a Range object as SourceData seems to be a known issue or several years that has not been fixed even in Office 2013. 

There are several stack overflow discussions where people who have passed Range objects to PivotCache.Create get a "type mismatch" error unexpectedly, even after the code ran successfully in previous testing. I encountered this where workbooks I had created with a Range object worked fine for weeks of testing, and then failed with "type mismatch" sporadically after the workbooks were put into circulation. Using a string as a workbook/cell range reference or a named range as a reference appeared to fix the problem. The official documentation should be updated, because passing a Range object here is unstable, yet the documentation specifies that is what should be passed.