Ad a "Known Issue" to warn MacOS developers on this method.
The method is broken on Mac and so far Microsoft has not fixed it. It triggers a runaway memory leakage, freezing the app and eventually crashing it. 

There should be a warning for developers trying to make their code compatible with Excel for Mac to not use this method until the bug is fixed. it is a longstanding issue with no ETA to be fixed.