Modifying code exemple
In the previous code exemple an out of bound access violation may occurs.
In the loop the r.Cells(n+1,1) instruction will try to reach the r.Rows.Count+1 column which is out of bound of the range r.
A Better test is to start à n = 2 and test with the previous cells. The code will works even if there is only 1 value in the range.