Change description of default value
(At least in Access 2010) the "Locked" property for form controls means that the control is locked if the property is set to True and unlocked if False. Unless there is some subtlety in the language that I'm missing, all descriptions of "Locked" are reversed. Only when Locked is False can a user edit, add or delete data.