Balanced coverage of FieldInfo and fixed table
Made coverage of FieldInfo more parallel for both Delimited and Fixed-width data, and moved some detail from initial table in Parameters section down to Remarks. Note that with only the names of the XlColumnDataType constants, and not the values, it's more difficult to understand the examples given. Also, the table for XlColumnDataType had not displayed correctly as formatted.