Added spaces before links
Added spaces before links - this is an issue in almost all of the VBA documentation. Looks like when it was converted, links lost the preceding space.
Same issue with italics under syntax - it is missing a space in almost all the documentation.

Is there anyway you could run a quick regex replace on everything?
/[^\s]\[/\ /
Would work for links.